trellomini/requirements.txt
Django
psycopg2
django-ratelimit
python-dotenv
djangorestframework